requirements/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
black==24.10.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
fastapi==0.115.6
flake8==7.1.1
grpcio==1.69.0rc1
grpcio-tools==1.69.0rc1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.0.1
idna==3.10
iniconfig==2.0.0
isort==5.13.2
jiter==0.8.2
Markdown==3.7
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
mypy==1.14.0
mypy-extensions==1.0.0
numpy==2.2.1
openai==1.58.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
portalocker==2.10.1
protobuf==5.29.2
pycodestyle==2.12.1
pydantic==2.10.4
pydantic_core==2.27.2
pyflakes==3.2.0
Pygments==2.18.0
pytest==8.3.4
python-dotenv==1.0.1
python-multipart==0.0.20
qdrant-client==1.12.2
regex==2024.11.6
requests==2.32.3
rfc3986==2.0.0
rich==13.9.4
setuptools==75.6.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.41.3
tabulate==0.9.0
tiktoken==0.8.0
tqdm==4.67.1
typer==0.15.1
types-Markdown==3.7.0.20241204
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
